Replace V1 with faster algorithm
Previously V1 was a verion of likelihood that did not use a loop. Using `@turbo` gives a 3x speedup and using the loop gives a 4x reduction in memory use. I am assuming that this will continue to scale as I add more calculations to the function. Therefore I'm remvoing the slow version of the function so I can evaluate changes in the new algorithm