Streamline calculations
Improve calculation of erosion and crustal area. Values are now created as a generic dictionary, iterated through, and converted to a NamedTuple. \n Comment out P flux calculations (need to be updated for new matched samples) and HDF5 code, since that creates `SIGKILL`, for some reason. Worth noting that the code won't run through because of this, since `rock_ersn` isn't calculated