Index using symbols
Slightly less efficient (because the string to index into the units vector is computed twice), but less likely to hit hard-to-find logic errors, because this is immune to any changes or mis-matches in vector element order. Potential to make this efficient again by checking that the unit name exists within the loop?