Remove unit check inefficiency
Doesn't calculate the string names for each unit twice anymore. Checks the units for each sample individually instead of all at once. The downside is that this will start the unit correction process before it throws an error, but becuase I've already run this through and I know it works that should be fine.