Remove key hardcoding
Previously erosion and continental area depended on hardcoded keys that made it so that calculating flux for only major elements would not have been possible. This removes that hardcoding so the keys of the erosion and crustal area tuples match the keys of the macro_cats NamedTuple. Also, added some minor reorganization for readability.