Update `likelihood` missing age assumption
Replace with the farthest possible age, rather than the negative sample age. This is because -4000 is much farther from +4000 than -3 is from -3. This will replace the missing age with 0 (if the rock is older than 1900 Myr) or 3800 (if the rock is younger than 1900 Myr), as these are the furthest possible ages from the sample that could be possible.